Move some configs from `sc-network-common` to `sc-network`
