+++ client/network/src/protocol.rs
@@ -16,8 +16,10 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::config;
+// Local dependencies
+use crate::{config, error};
 
+// External dependencies
 use bytes::Bytes;
 use codec::{DecodeAll, Encode};
 use libp2p::{
@@ -29,26 +31,30 @@ use libp2p::{
 	Multiaddr, PeerId,
 };
 use log::{debug, error, warn};
-use message::{generic::Message as GenericMessage, Message};
-use notifications::{Notifications, NotificationsOut};
+
+// Substrate dependencies
 use sc_network_common::{
 	config::NonReservedPeerMode,
-	error,
 	protocol::{role::Roles, ProtocolName},
 	sync::message::BlockAnnouncesHandshake,
 };
 use sp_runtime::traits::Block as BlockT;
+
+// `std` dependencies
 use std::{
 	collections::{HashMap, HashSet, VecDeque},
 	iter,
 	task::Poll,
 };
 
+use message::{generic::Message as GenericMessage, Message};
+use notifications::{Notifications, NotificationsOut};
+
+pub use notifications::{NotificationsSink, NotifsHandlerError, Ready};
+
 mod notifications;
 
 pub mod message;
-
-pub use notifications::{NotificationsSink, NotifsHandlerError, Ready};
 
 /// Maximum size used for notifications in the block announce and transaction protocols.
 // Must be equal to `max(MAX_BLOCK_ANNOUNCE_SIZE, MAX_TRANSACTIONS_SIZE)`.
@@ -93,7 +99,7 @@ impl<B: BlockT> Protocol<B> {
 	pub fn new(
 		roles: Roles,
 		network_config: &config::NetworkConfiguration,
-		block_announces_protocol: sc_network_common::config::NonDefaultSetConfig,
+		block_announces_protocol: config::NonDefaultSetConfig,
 	) -> error::Result<(Self, sc_peerset::PeersetHandle, Vec<(PeerId, Multiaddr)>)> {
 		let mut known_addresses = Vec::new();
