+++ client/service/src/config.rs
@@ -22,14 +22,11 @@ pub use sc_client_api::execution_extensions::{ExecutionStrategies, ExecutionStra
 pub use sc_client_db::{BlocksPruning, Database, DatabaseSource, PruningMode};
 pub use sc_executor::{WasmExecutionMethod, WasmtimeInstantiationStrategy};
 pub use sc_network::{
-	config::{NetworkConfiguration, Role},
+	config::{NetworkConfiguration, NonDefaultSetConfig, Role},
 	Multiaddr,
 };
 pub use sc_network_common::{
-	config::{
-		MultiaddrWithPeerId, NodeKeyConfig, NonDefaultSetConfig, ProtocolId, SetConfig,
-		TransportConfig,
-	},
+	config::{MultiaddrWithPeerId, NodeKeyConfig, ProtocolId, SetConfig, TransportConfig},
 	request_responses::{
 		IncomingRequest, OutgoingResponse, ProtocolConfig as RequestResponseConfig,
 	},
